Use --input_types and --num_classes instead of --datasets
I don't fully remember my rationale for this. I think I was about to use
args.dataset to determine how to compute frame offsets (which will be in
the very next commit), and I was about to add a dummy dataset for
charades just so that the JsonDatasetLoader would recognize it and know
how to compute frame offsets on it.

Besides, it doesn't really make sense to specify, e.g.,
'davis_rgb_moving_test' when running on an arbitrary internet video.